fix: 페이지네이션 초기화 문제 수정 및 로컬스토리지 저장 #89
