+++ src/components/main/Select/Select.tsx
@@ -6,9 +6,10 @@ import { SelectOptionList } from '../../common/Select/SelectOptionList/SelectOpt
 
 interface SelectProps {
   onSubmitOption: () => void;
+  setCurrentPageNumber: (currentPageNumber: number) => void;
 }
 
-const Select = ({ onSubmitOption }: SelectProps) => {
+const Select = ({ onSubmitOption, setCurrentPageNumber }: SelectProps) => {
   const optionListRef = useRef<HTMLUListElement>(null);
   const countryTabRef = useRef<HTMLDivElement>(null);
   const regionTabRef = useRef<HTMLDivElement>(null);
@@ -31,6 +32,7 @@ const Select = ({ onSubmitOption }: SelectProps) => {
       <SelectOptionList
         isOpen={isOpen}
         setIsOpen={setIsOpen}
+        setCurrentPageNumber={setCurrentPageNumber}
         optionListRef={optionListRef}
       />
     </SelectContainer>
